[batch] Fix deadlock with exceeded shares
