+++ batch/batch/driver/instance_pool.py
@@ -1,4 +1,3 @@
-from collections import deque
 import os
 import secrets
 import random
@@ -20,48 +19,13 @@ from ..batch_configuration import DEFAULT_NAMESPACE, PROJECT, \
 
 from .instance import Instance
 from ..worker_config import WorkerConfig
+from ..utils import WindowFractionCounter
 
 log = logging.getLogger('instance_pool')
 
 BATCH_WORKER_IMAGE = os.environ['HAIL_BATCH_WORKER_IMAGE']
 
 log.info(f'BATCH_WORKER_IMAGE {BATCH_WORKER_IMAGE}')
-
-
-class WindowFractionCounter:
-    def __init__(self, window_size: int):
-        self._window_size = window_size
-        self._q = deque()
-        self._n_true = 0
-        self._seen = set()
-
-    def push(self, key: str, x: bool):
-        self.assert_valid()
-        if key in self._seen:
-            return
-        while len(self._q) >= self._window_size:
-            old_key, old = self._q.popleft()
-            self._seen.remove(old_key)
-            if old:
-                self._n_true -= 1
-        self._q.append((key, x))
-        self._seen.add(key)
-        if x:
-            self._n_true += 1
-        self.assert_valid()
-
-    def fraction(self) -> float:
-        self.assert_valid()
-        # (1, 1) prior
-        return (self._n_true + 1) / (len(self._q) + 2)
-
-    def __repr__(self):
-        self.assert_valid()
-        return f'{self._n_true}/{len(self._q)}'
-
-    def assert_valid(self):
-        assert len(self._q) <= self._window_size
-        assert 0 <= self._n_true <= self._window_size
 
 
 class ZoneSuccessRate:
